** Fix for bug #544031
2008-07-25  Milan Crha  <mcrha@redhat.com>

	** Fix for bug #544031

	* camel/camel-vee-folder.c: (vee_folder_remove_folder):
	* camel/camel-folder-summary.c: (camel_folder_summary_migrate_infos),
	(content_info_to_db): Compiler warnings cleanup.

	* camel/providers/local/camel-mbox-summary.c: (summary_header_to_db):
	Compiler warning cleanup.


svn path=/trunk/; revision=9191

+++ camel/providers/local/camel-mbox-summary.c
@@ -303,7 +303,7 @@ summary_header_to_db (CamelFolderSummary *s, CamelException *ex)
 	fir = ((CamelFolderSummaryClass *)camel_mbox_summary_parent)->summary_header_to_db (s, ex);
 	if (fir) {
 		tmp = fir->bdata;
-		fir->bdata = g_strdup_printf ("%s %d %d", tmp ? tmp : "", CAMEL_MBOX_SUMMARY_VERSION, mbs->folder_size);
+		fir->bdata = g_strdup_printf ("%s %d %d", tmp ? tmp : "", CAMEL_MBOX_SUMMARY_VERSION, (int) mbs->folder_size);
 		g_free (tmp);
 	}
